Extract computeSourceStatus to separate file
