+++ packages/sources/src/addSourceWizard/FinalWizard.js
@@ -11,30 +11,7 @@ import TimeoutStep from './steps/TimeoutStep';
 
 import { WIZARD_DESCRIPTION, WIZARD_TITLE } from '../utilities/stringConstants';
 import { getSourcesApi } from '../api';
-
-export const getSourceStatus = source => {
-    const appStatus = source.applications?.[0]?.availability_status;
-    const endpointStatus = source.endpoint?.[0]?.availability_status;
-    const statuses = [ appStatus, endpointStatus ];
-
-    if (statuses.includes('unavailable')) {
-        return 'unavailable';
-    }
-
-    if (
-        (appStatus === 'available' && !source.endpoint?.[0])
-        || (endpointStatus === 'available' && !source.applications?.[0])
-        || (appStatus === 'available' && endpointStatus === 'available')
-    ) {
-        return 'available';
-    }
-
-    if (source.applications?.[0] || source.endpoint?.[0]) {
-        return 'timeout';
-    }
-
-    return 'finished';
-};
+import computeSourceStatus from '../utilities/computeSourceStatus';
 
 const FinalWizard = ({
     afterSubmit,
@@ -86,7 +63,7 @@ const FinalWizard = ({
             })}
         />;
     } else if (isFinished) {
-        switch (getSourceStatus(createdSource)) {
+        switch (computeSourceStatus(createdSource)) {
             case 'unavailable':
                 step = <ErroredStep
                     onClose={ afterSubmit }
