+++ packages/sources/src/tests/addSourceWizard/finalWizard.test.js
@@ -249,100 +249,4 @@ describe('Final wizard', () => {
         expect(wrapper.find(FinishedStep)).toHaveLength(1);
         expect(wrapper.find(EmptyState).find(Title).text()).toEqual('Configuration successful');
     });
-
-    describe('getSourceStatus', () => {
-        let source;
-
-        it('endpoint is unavailable', () => {
-            source = {
-                endpoint: [{
-                    availability_status: 'unavailable'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('unavailable');
-        });
-
-        it('application is unavailable', () => {
-            source = {
-                applications: [{
-                    availability_status: 'unavailable'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('unavailable');
-        });
-
-        it('endpoint is available', () => {
-            source = {
-                endpoint: [{
-                    availability_status: 'available'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('available');
-        });
-
-        it('endpoint is available && applications is timeouted', () => {
-            source = {
-                endpoint: [{
-                    availability_status: 'available'
-                }],
-                applications: [{
-                    availability_status: 'null'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('timeout');
-        });
-
-        it('application is available', () => {
-            source = {
-                applications: [{
-                    availability_status: 'available'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('available');
-        });
-
-        it('endpoint && application are available', () => {
-            source = {
-                endpoint: [{
-                    availability_status: 'available'
-                }],
-                applications: [{
-                    availability_status: 'available'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('available');
-        });
-
-        it('endpoint is timeouted', () => {
-            source = {
-                endpoint: [{
-                    availability_status: null
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('timeout');
-        });
-
-        it('application is timeouted', () => {
-            source = {
-                applications: [{
-                    availability_status: 'null'
-                }]
-            };
-
-            expect(getSourceStatus(source)).toEqual('timeout');
-        });
-
-        it('returns default', () => {
-            source = {};
-
-            expect(getSourceStatus(source)).toEqual('finished');
-        });
-    });
 });
